pyproject.toml
[project]
name = "litelitellm"
version = "0.1.0"
description = "Lightweight Anthropic Messages API proxy with LiteLLM-style config — load your middleware from any project"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
litelitellm = "litelitellm.__main__:main"

[project.optional-dependencies]
dev = ["flake8>=7.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501"]
exclude = [".venv", "__pycache__", "build", "dist"]
per-file-ignores = ["__init__.py:F401"]
